server/requirements.txt
fastapi
"uvicorn[standard]"
mysql-connector-python
python-dotenv
python-socketio
pyjwt
redis
ultralyticsplus
opencv-python
numpy

serial
pynema2
geopy